index.js and and shortened functions
- created common functions in render/api to avoid duplicated code
- moved code from index.js to separate module (commented original code just in case)
