+++ src/index.js
@@ -1,33 +1,34 @@
 import './sass/main.scss';
 import './js-modules/trendingMoviesSwitcher.js';
 import './js-modules/auth.js';
-import { getGenres } from './js-modules/apiService.js';
+// import { getGenres } from './js-modules/apiService.js';
 import './js-modules/galleryMarkup.js';
 import './js-modules/themeSwitcher.js';
-import { renderTrendingFilms } from './js-modules/render-service.js';
-import variables from './js-modules/variables.js';
+// import { renderTrendingFilms } from './js-modules/render-service.js';
+// import variables from './js-modules/variables.js';
 import './js-modules/modal-film-card';
 import './js-modules/modal-footer.js'; 
-import Pages from './js-modules/pagination.js'; 
-import initUpArrow from './js-modules/up-arrow.js';
+// import Pages from './js-modules/pagination.js'; 
+// import initUpArrow from './js-modules/up-arrow.js';
 import './js-modules/btnListener.js';
 import './js-modules/isOpenGenre';
 import './js-modules/btnListener.js';
+import "./js-modules/initialMarkup.js";
 
 // Рендер панели с пагинацией и первой страницы с трендовыми фильмами
-export const pagination = new Pages('.pagination');
-export const initMainMarkup = async function () {
-  const totalPages = await renderTrendingFilms(variables.filmGrid, variables.preloader);
-  pagination.moveToPage(1, totalPages);
-};
-// Полноценная пагинация по трендовым фильмам
-export const updateTrendingMarkup = async function () {
-  const newPage = pagination.page;
-  await renderTrendingFilms(variables.filmGrid, variables.preloader, newPage);
-};
-pagination.listen(updateTrendingMarkup);
-// *********************************************************************
+// export const pagination = new Pages('.pagination');
+// export const initMainMarkup = async function () {
+//   const totalPages = await renderTrendingFilms(variables.filmGrid, variables.preloader);
+//   pagination.moveToPage(1, totalPages);
+// };
+// // Полноценная пагинация по трендовым фильмам
+// export const updateTrendingMarkup = async function () {
+//   const newPage = pagination.page;
+//   await renderTrendingFilms(variables.filmGrid, variables.preloader, newPage);
+// };
+// pagination.listen(updateTrendingMarkup);
+// // *********************************************************************
 
-getGenres();
-initMainMarkup();
-initUpArrow();
+// getGenres();
+// initMainMarkup();
+// initUpArrow();
