+++ src/js-modules/galleryMarkup.js
@@ -1,9 +1,10 @@
 import trendingFilmsMarkupFc from "../hbs-templates/trending-films.hbs";
 
-import { pagination } from "../index";
+// import { pagination } from "../index";
 import api from './apiService.js'
 import './btnListener.js';
 import variables from "./variables.js";
+import { pagination } from './initialMarkup.js';
